+++ bc/service_directory/services.py
@@ -13,7 +13,7 @@ def format_service_detail_page_url(
     category_filters: Optional[Sequence[str]] = None,
     collection_filter: Optional[str] = None,
     frontend_url: str,
-):
+) -> str:
     query_dict = QueryDict(mutable=True)
 
     if collection_filter is not None:
@@ -23,10 +23,7 @@ def format_service_detail_page_url(
         for category in category_filters:
             query_dict.appendlist("categories", category)
 
-    if frontend_url.endswith("/"):
-        frontend_url = frontend_url[:-1]
-
-    url = f"{frontend_url}/service/{quote_plus(str(service.id))}"
+    url = f"{frontend_url.rstrip('/')}/service/{quote_plus(str(service.id))}"
 
     if query_dict:
         url += f"?{query_dict.urlencode()}"
@@ -38,7 +35,7 @@ def format_services_listing_page_url(
     category_filters: Optional[Sequence[str]] = None,
     collection_filter: Optional[str] = None,
     frontend_url: str,
-):
+) -> str:
     query_dict = QueryDict(mutable=True)
 
     if collection_filter is not None: